+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.3.0] - 2015-11-17
 ### Added
 - datepicker with date range and filtering _(currently in angular-material 1.0_RC4)_
 - Tests of formlyMaterial provider
@@ -42,7 +42,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.1 - 2015-11-06
 
-[Unreleased]: https://github.com/kamilkisiela/meteor-angular-formly-templates-material/compare/v0.2.0...HEAD
+[0.3.0]: https://github.com/kamilkisiela/meteor-angular-formly-templates-material/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/kamilkisiela/meteor-angular-formly-templates-material/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/kamilkisiela/meteor-angular-formly-templates-material/compare/v0.0.4...v0.1.0
 [0.0.4]: https://github.com/kamilkisiela/meteor-angular-formly-templates-material/compare/v0.0.3...v0.0.4